ai-service/requirements.txt
fastapi
uvicorn
pydantic
transformers==4.41.2
minio
mammoth
pytesseract
Pillow
PyMuPDF

